refactor(crypto-helper): asn1: use reference counting to reduce data cloning.
Previously, for every hex byte node, NodeOptions component, copy operation,
and so on, the app has cloned entire node bytes. This patch introduces
the `RcSlice` structure. This structure represents the slice with multiple
owners. But actually, it's just a `Vec` in the `Rc` plus range.

These changes affect the hex viewer and `NodeOptions` components the most.
We don't want to cause a big refactoring for now.

We have better performance and memory efficiency. `Vec`s cloning now is
not a bottleneck. But the noticeable freezes are still present.